updated styling and added responsiveness, got footers to stick to the bottom

+++ src/components/PetPage.jsx
@@ -22,10 +22,14 @@ export function PetPage() {
   const history = useHistory()
 
   async function deletePet() {
-    const response = await axios.delete(
-      `https://tamagotchi-api-bradygrapentine.herokuapp.com/api/Pets/${id}`
-    )
-    if (response.status == 200) {
+    if (!pet.isDead) {
+      const response = await axios.delete(
+        `https://tamagotchi-api-bradygrapentine.herokuapp.com/api/Pets/${id}`
+      )
+      if (response.status == 200) {
+        history.push('/')
+      }
+    } else {
       history.push('/')
     }
   }
@@ -59,20 +63,22 @@ export function PetPage() {
   }
   async function renamePet(event) {
     event.preventDefault()
-    const response = await axios.put(
-      `https://tamagotchi-api-bradygrapentine.herokuapp.com/api/Pets/${id}`,
-      {
-        id: pet.id,
-        name: newName,
-        birthday: pet.birthday,
-        hungerLevel: pet.hungerLevel,
-        happinessLevel: pet.happinessLevel,
-        lastInteractedWithDate: pet.lastInteractedWithDate,
-        isDead: pet.isDead,
+    if (!pet.isDead) {
+      const response = await axios.put(
+        `https://tamagotchi-api-bradygrapentine.herokuapp.com/api/Pets/${id}`,
+        {
+          id: pet.id,
+          name: newName,
+          birthday: pet.birthday,
+          hungerLevel: pet.hungerLevel,
+          happinessLevel: pet.happinessLevel,
+          lastInteractedWithDate: pet.lastInteractedWithDate,
+          isDead: pet.isDead,
+        }
+      )
+      if (response.status == 200) {
+        getPet()
       }
-    )
-    if (response.status == 200) {
-      getPet()
     }
     setNewName('')
   }
@@ -103,43 +109,67 @@ export function PetPage() {
   return (
     <>
       <Link to="/">
-        <h2 className="petPage">Interact With Your Katagotchis!</h2>
+        <h3 className="petPage">Interact With Your Katagotchis!</h3>
       </Link>
       {pet.id === '' ? (
         <p>Loading...</p>
       ) : (
         <>
-          <ul className="petPage">
-            <li className="name">
-              Name: <p>{pet.name}</p>
-            </li>
-            <li className="bday">
-              Birthday: <p>{formatDate(pet.birthday)}</p>
-            </li>
-            <li className="hunger">
-              Hunger Level: <p>{pet.hungerLevel}</p>
-            </li>
-            <li className="happiness">
-              Happiness Level: <p>{pet.happinessLevel}</p>
-            </li>
-            <li className="isDead">
-              Status: <p>{pet.isDead ? 'Dead' : 'Alive'}</p>
-            </li>
-            <li className="lastInteraction">
-              Last Interaction: <p>{formatDate(pet.lastInteractedWithDate)}</p>
-            </li>
+          <ul className="deleteAndRename">
+            <form className="rename" onSubmit={renamePet}>
+              <label>Rename Katagotchi:</label>
+              <input
+                type="text"
+                placeholder="Enter New Name"
+                value={newName}
+                onChange={event => {
+                  setNewName(event.target.value)
+                }}
+              />
+            </form>
+            <button className="delete" onClick={deletePet}>
+              Delete Katagotchi
+            </button>
+          </ul>
+          <div className="petPage">
+            <ul className="petPage">
+              <li className="name">
+                Name: <p>{pet.name}</p>
+              </li>
+              <li className="bday">
+                Birthday: <p>{formatDate(pet.birthday)}</p>
+              </li>
+              <li className="hunger">
+                Hunger Level: <p>{pet.hungerLevel}</p>
+              </li>
+              <li className="happiness">
+                Happiness Level: <p>{pet.happinessLevel}</p>
+              </li>
+              <li className="isDead">
+                Status: <p>{pet.isDead ? 'Dead' : 'Alive'}</p>
+              </li>
+              <li className="lastInteraction">
+                Last Interaction:{' '}
+                <p>{formatDate(pet.lastInteractedWithDate)}</p>
+              </li>
+            </ul>
             {pet.isDead ? (
               <p>
                 Well then...just because they're made of data, doesn't mean they
                 don't need to eat...
               </p>
             ) : (
-              <>
+              <ul className="petInteractions">
                 <li className="interactions">
-                  <img
-                    src="https://placekeanu.com/240/280/y"
-                    alt="No Image Available"
-                  />
+                  <a
+                    className="katagotchiImage"
+                    href="https://placekeanu.com/240/280/y"
+                  >
+                    <img
+                      src="https://placekeanu.com/240/280/y"
+                      alt="No Image Available"
+                    />
+                  </a>
                   <button className="play" onClick={playWithPet}>
                     Play With Katagotchi
                   </button>
@@ -150,32 +180,17 @@ export function PetPage() {
                     Feed Katagotchi
                   </button>
                 </li>
-                <li className="delete">
-                  <form className="rename" onSubmit={renamePet}>
-                    <label>Rename Katagotchi:</label>
-                    <input
-                      type="text"
-                      placeholder="Enter New Name"
-                      value={newName}
-                      onChange={event => {
-                        setNewName(event.target.value)
-                      }}
-                    />
-                  </form>
-                </li>
-                <li className="delete">
-                  <button className="delete" onClick={deletePet}>
-                    Delete Katagotchi
-                  </button>
-                </li>
-              </>
+              </ul>
             )}
-          </ul>
-          <Link to="/">
-            <button className="petPageToHome">Back to Katagotchi List</button>
-          </Link>
+          </div>
         </>
       )}
+      <footer>
+        {' '}
+        <Link to="/">
+          <button className="petPageToHome">Back to Katagotchi List</button>
+        </Link>
+      </footer>
     </>
   )
 }
